pyproject.toml
[tool.poetry]
name = "aquaapi"
version = "0.1.0"
description = ""
authors = ["Jakub Wilk <wilkjakub64@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.0.2"
Flask-PyMongo = "^2.3.0"
PyJWT = "^2.3.0"
dnspython = "^2.1.0"
bcrypt = "^3.2.0"
Flask-Cors = "^3.0.10"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
